Allow retention policy to be specified for write
According to the InfluxDB docs, when writing, the retention policy can
be specified with a URL parameter named `rp`:

https://docs.influxdata.com/influxdb/v1.2/guides/writing_data/#writing-data-using-the-http-api

I was thinking that perhaps the retention policy name could be specified
in the Series definition.  Then multiple points could be written, each
of which could have its own retention policy. In communicating with
Influx support, I heard back this:

"There is currently no way to specify the database or retention policy
directly in the point at this time. Although a per-point retention
policy option is already a proposed feature for the next version of the
line protocol format."

So, we can specify the retention policy name as an option to the `write`
function alongside other options like `precision`.